Refactor: Extract API calls to a separate service file
